Changing the guildExp data type from integer to double

+++ hypixel-api-example/src/main/java/net/hypixel/api/example/GetGuildExample.java
@@ -180,7 +180,7 @@ public class GetGuildExample {
          * Then we loop through each game and see how much experience the guild's earned from it.
          */
         for (GameType game : GameType.values()) {
-            int experienceForGame = guild.getExperienceForGame(game);
+            double experienceForGame = guild.getExperienceForGame(game);
             System.out.println("\t" + game.getName() + ": " + experienceForGame);
         }
     }
